use unsigned int for angle

+++ physarumgen/include/agent.h
@@ -5,9 +5,9 @@
 class c_agent
 {
     public:
-        c_agent(unsigned int pos_x, unsigned int pos_y, float angle);
+        c_agent(unsigned int pos_x, unsigned int pos_y, unsigned int angle);
         void print_info() const;
         s_position m_position;
-        float m_angle;
+        unsigned int m_angle;
     private:
 };
